+++ src/components/common/Modal/index.ts
@@ -6,3 +6,6 @@ export type { AlertModalProps } from './AlertModal'
 
 export { ConfirmModal } from './ConfirmModal'
 export type { ConfirmModalProps } from './ConfirmModal'
+
+export { RestoreModal } from './RestoreModal'
+export type { RestoreModalProps } from './RestoreModal'
